docs(Tabs): move to components folder
