+++ src/components/Tabs/docs/TabsExtendedExample.st.css
@@ -1,5 +1,5 @@
 :import {
-    -st-from: "../../src/components/Tabs/Tabs.st.css";
+    -st-from: "../Tabs.st.css";
     -st-default: TPATabs;
 }
