+++ src/components/Tabs/docs/TabsExtendedExample.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {Tabs, TabsProps} from '../../src/components/Tabs';
+import {Tabs, TabsProps} from '../Tabs';
 import extendedStyles from './TabsExtendedExample.st.css';
 
 export const TabsExtendedExample: React.SFC<TabsProps> =  (props) => <Tabs {...props} {...extendedStyles('root', {}, props)}/>;
